first pass at refactor
