+++ anvilfs/tests/testconfig.py
@@ -1,0 +1,9 @@
+# python-compatible syntax to define required vars:
+# 
+# namespace_name: <your AnVIL Google project name>
+#    NOTE: must have google cloud sdk configured to this project
+# workspace_name: <the target AnVIL workspace name>
+config = {
+    "namespace_name": "<some project>",
+    "workspace_name": "<some workspace>"
+}
