+++ anvilfs/tests/unit.py
@@ -1,0 +1,8 @@
+from ..bucket import WorkspaceBucket
+
+# test bucket
+def test_bucket(bucket_name):
+    bucket = Workspacebucket(bucket_name)
+
+def run_all():
+    pass
