+++ run_tests.sh
@@ -1,1 +1,2 @@
+#!/usr/bin/env bash
 python -m anvilfs.tests.run_tests
